Guiding Center Dynamics in 4D

Guiding centre dynamics is a reduced version of charged particle dynamics, where the motion of the particle in a strong magnetic field $B$ is reduced to the motion of the guiding centre, that is the centre of the gyro motion of the particle about a magnetic field line. The dynamics of the guiding centre can be described in terms of only four coordinates (as compared to six for the full motion of the charged particle), the position of the guiding centre $r = (x,y,z)$ and the parallel velocity $u$, where parallel refers to the direction of the magnetic field.

Lagrangian Formulation

The simplest form of the guiding centre equations can be obtained from the Lagrangian

\[\begin{aligned} L &= (A (r) + u b (r)) \cdot \dot{r} - H(r,u) , & H &= \tfrac{1}{2} u^{2} + \mu \vert B (r) \vert , \end{aligned}\]

where $b = B / \vert B \vert$ is the unit vector of the magnetic field $B = \nabla \times A$ with $A$ the magnetic vector potential and $\mu$ is the magnetic moment. The Euler-Lagrange equations are computed as

\[\begin{aligned} \nabla \vartheta^{T} ( r(t), u(t) ) \cdot \dot{r} (t) - \dot{\vartheta} ( r(t), u(t) ) &= \nabla H ( r(t), u(t) ) , \\ b ( r(t) ) \cdot \dot{r} (t) &= u (t) , \end{aligned}\]

with $\vartheta(r,u) = A (r) + u \, b (r)$ and the gradient $\nabla$ denoting the derivative with respect to $r$.

Hamiltonian Formulation

Computing the time derivative of $\vartheta$, the Euler-Lagrange equations can be rewritten in an explicit form as

\[\begin{aligned} \dot{r} (t) &= \dfrac{u (t) \, \beta (r (t))}{b (r (t)) \cdot \beta (r (t))} + \dfrac{B (r (t))}{B (r (t)) \cdot \beta (r (t))} \times \nabla H (r (t),u (t)) , \\ \dot{u} (t) &= - \dfrac{\beta (r (t))}{b (r (t)) \cdot \beta (r (t))} \cdot \nabla H (r (t), u (t)) , \end{aligned}\]

where $\beta = \nabla \times \vartheta$. This constitutes a noncanonical Hamiltonian system of the form

\[\dot{q} = \Omega^{-T} (q) \nabla H(q) ,\]

with $q = (x,y,z,u)$ and the symplectic matrix $\Omega$ given by

\[\Omega_{ij} = \dfrac{\partial \vartheta_{j}}{\partial q^{i}} - \dfrac{\partial \vartheta_{i}}{\partial q^{j}} .\]

Modules

ChargedParticleDynamics.GuidingCenter4d.SymmetricField — Module

First and second Poincaré invariant for a guiding center particle in an axisymmetric magnetic field of the form $B(x,y,z) = B_0 (1 + x^2 + y^2) e_z$.

The loop for the first invariant is initialized by

\[q (\tau) = \begin{pmatrix} r_x \cos (2\pi \tau) \\ r_y \sin (2\pi \tau) \\ z_0 + z_1 \sin (2\pi \tau) \\ u_0 + u_1 \cos (2\pi \tau) \\ \end{pmatrix}\]

with parameters

\[B_0 = 1, \quad r_x = 0.5, \quad r_y = 0.3, \quad z_0 = 0.0, \quad z_1 = 0.1, \quad u_0 = 0.5, \quad u_1 = 0.05, \quad \mu = 0.01 .\]

The surface for the second invariant is initialized by

\[q (\tau) = \begin{pmatrix} r_0 (\sigma - 0.5) \\ r_0 (\tau - 0.5) \\ z_0 + z_1 \cos (2\pi \sigma) \cos (2\pi \tau) \\ u_0 + u_1 \sin (2\pi \sigma) \sin (2\pi \tau) \\ \end{pmatrix}\]

with parameters

\[B_0 = 1, \quad r_0 = 0.5, \quad z_0 = 0.0, \quad z_1 = 0.1, \quad u_0 = 0.5, \quad u_1 = 0.01, \quad \mu = 0.01 .\]

ChargedParticleDynamics.GuidingCenter4d.ThetaPinchField — Module

First Poincaré invariant for a guiding center particle in an θ-pinch magnetic field of the form $B(x,y,z) = B_0 \, e_z$.

The loop for the first Poincaré invariant is initialized by

\[q (\tau) = \begin{pmatrix} r_x \cos (2\pi \tau) \\ y_0 \\ r_z \sin (2\pi \tau) \\ \end{pmatrix}\]

with parameters

\[B_0 = 1, \quad r_x = 0.5, \quad r_z = 0.3, \quad y_0 = 0.0, \quad u_0 = 0.5, \quad \mu = 2.5 \times 10^{-6} .\]

ChargedParticleDynamics.GuidingCenter4d.TokamakSmallCartesian — Module

Analytic axisymmetric small tokamak equilibrium in cartesian coordinates.

ChargedParticleDynamics.GuidingCenter4d.TokamakSmallCylindrical — Module

Analytic axisymmetric small tokamak equilibrium in cylindrical coordinates.

ChargedParticleDynamics.GuidingCenter4d.TokamakSmallToroidal — Module

Analytic axisymmetric small tokamak equilibrium in circular coordinates.

ChargedParticleDynamics.GuidingCenter4d.TokamakMediumCartesian — Module

Analytic axisymmetric medium-size tokamak equilibrium in cartesian coordinates.

ChargedParticleDynamics.GuidingCenter4d.TokamakMediumCylindrical — Module

Analytic axisymmetric medium-size tokamak equilibrium in cartesian coordinates.

ChargedParticleDynamics.GuidingCenter4d.TokamakIterCylindrical — Module

Analytic ITER-like Solov'ev equilibrium with X-point.

ChargedParticleDynamics.GuidingCenter4d.SolovevIter — Module

Analytic ITER-like Solov'ev equilibrium.

ChargedParticleDynamics.GuidingCenter4d.SolovevIterXpoint — Module

Analytic ITER-like Solov'ev equilibrium with X-point.